remote_support/requirements.txt
# Dependencias para CompuEasys Remote Support
pillow>=10.0.0
pyautogui>=0.9.54
requests>=2.31.0
